Reorganize factories by family
