+++ C07/AbstractFactory/src/Vehicles/HighEnd/HighEndBike.cs
@@ -1,3 +1,3 @@
-﻿namespace Vehicles.Models;
+﻿namespace Vehicles.HighEnd;
 
 public class HighEndBike : IBike { }
