+++ C07/AbstractFactory/src/Vehicles/HighEnd/HighEndCar.cs
@@ -1,3 +1,3 @@
-﻿namespace Vehicles.Models;
+﻿namespace Vehicles.HighEnd;
 
 public class HighEndCar : ICar { }
