+++ C07/AbstractFactory/src/Vehicles/IBike.cs
@@ -1,0 +1,3 @@
+﻿namespace Vehicles;
+
+public interface IBike { }
