+++ C07/AbstractFactory/src/Vehicles/ICar.cs
@@ -1,0 +1,3 @@
+﻿namespace Vehicles;
+
+public interface ICar { }
