+++ C07/AbstractFactory/src/Vehicles/IVehicleFactory.cs
@@ -1,6 +1,4 @@
-﻿using Vehicles.Models;
-
-namespace Vehicles;
+﻿namespace Vehicles;
 
 public interface IVehicleFactory
 {
