+++ C07/AbstractFactory/src/Vehicles/LowEnd/LowEndBike.cs
@@ -1,3 +1,3 @@
-﻿namespace Vehicles.Models;
+﻿namespace Vehicles.LowEnd;
 
 public class LowEndBike : IBike { }
