+++ C07/AbstractFactory/src/Vehicles/LowEnd/LowEndCar.cs
@@ -1,3 +1,3 @@
-﻿namespace Vehicles.Models;
+﻿namespace Vehicles.LowEnd;
 
 public class LowEndCar : ICar { }
