+++ C07/AbstractFactory/src/Vehicles/MidRange/MidRangeCar.cs
@@ -1,3 +1,3 @@
-﻿namespace Vehicles.Models;
+﻿namespace Vehicles.MidRange;
 
 public class MidRangeCar : ICar { }
