--- C07/AbstractFactory/src/Vehicles/Models/IBike.cs
@@ -1,3 +1,0 @@
-﻿namespace Vehicles.Models;
-
-public interface IBike { }
