+++ C07/AbstractFactory/test/Vehicles.Tests/AbstractFactoryTest.cs
@@ -1,11 +1,6 @@
-using Vehicles.Models;
 using Xunit;
 
 namespace Vehicles;
-
-public class LowEndVehicleFactoryTest : BaseAbstractFactoryTest<LowEndVehicleFactory, LowEndCar, LowEndBike> { }
-public class MidRangeVehicleFactoryTest : BaseAbstractFactoryTest<MidRangeVehicleFactory, MidRangeCar, MidRangeBike> { }
-public class HighEndVehicleFactoryTest : BaseAbstractFactoryTest<HighEndVehicleFactory, HighEndCar, HighEndBike> { }
 
 public abstract class BaseAbstractFactoryTest<TConcreteFactory, TExpectedCar, TExpectedBike>
     where TConcreteFactory : IVehicleFactory, new()
