+++ C07/AbstractFactory/test/Vehicles.Tests/HighEnd/HighEndVehicleFactoryTest.cs
@@ -1,0 +1,3 @@
+﻿namespace Vehicles.HighEnd;
+
+public class HighEndVehicleFactoryTest : BaseAbstractFactoryTest<HighEndVehicleFactory, HighEndCar, HighEndBike> { }
