+++ C07/AbstractFactory/test/Vehicles.Tests/LowEnd/LowEndVehicleFactoryTest.cs
@@ -1,0 +1,3 @@
+﻿namespace Vehicles.LowEnd;
+
+public class LowEndVehicleFactoryTest : BaseAbstractFactoryTest<LowEndVehicleFactory, LowEndCar, LowEndBike> { }
