+++ C07/AbstractFactory/test/Vehicles.Tests/MidRange/MidRangeVehicleFactoryTest.cs
@@ -1,0 +1,3 @@
+﻿namespace Vehicles.MidRange;
+
+public class MidRangeVehicleFactoryTest : BaseAbstractFactoryTest<MidRangeVehicleFactory, MidRangeCar, MidRangeBike> { }
